cloud-function/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.0
requests==2.31.0
supabase==2.8.0
google-generativeai